speedup things, allow brake on digital output
- speed up communication via parallel commands, switchable with
  asynSetOption <port> <addr> "pollMethod" <mode>
  (<mode>: serial, axis-parallel, ctrl-parallel, <addr>==0: controller)
- add brake support with digital output with new command "phytronBrakeOutput"
  or additional EPICS-PVs
- allow to automatically enable/disable motor output
- allow autosave-ing "DVAL"
- generate and check messages with CRC

The default settings behave like the previous version of the support module,
but message CRC is generated and checked. It uses the serial poll method
(one command, one reply), no brake, no stage output en-/disabling.

It is worth using a parallel poll method if your controller has the firmware
v1.1.12 or later: the normal round trip time is 10ms, the old serial method
uses four commands, which is about 40ms multiplied with the number of motors.
For a 10 motor phyMOTION it takes 400ms for a single poll, a full stacked
phyMOTION with 64 motors needs about 2.5s! The parallel method uses mostly
one round trip time with a longer command line and reply (up to 1000 bytes
for a command line as suggested by Phytron, up to 2000 bytes per reply).