use one database "Phytron_motor.db" only with zero defaults for MRES,SREV,UREV
A harmonic solution was presented by Mark Rivers, who pointed out that setting
zero defaults for MRES,SREV,UREV requires to set MRES or UREV (and optional
SREV). This needs only one file and this allows the user to decide, which
field(s) should be used to initialize the motor.